refactor: move providers to separate module
